Add authentication package
We'd like to be generating some credentials on the fly, and expire them
after a short while for security. Redis happens to be a perfect engine
for our desired behaviour.

These credentials, will then be used in the Web server that will
generate and distribute them to users, socks5 and http proxy to
authenticate with the server.