Define main package
We'd like to have an app that is capable to be executed on a remote
server and serve different purposes, for instance a web server or a
socks5 proxy.

This could be separate subcommands, making a single binary a source of
truth and reduce amount of inconsistency.

Using kingpin, as it allows for CLI setup with subcommands, flags and
environment variables in the same time. It allows us to have defaults
for our options which may make our lives easier down the line.