Add web server
The goal is to have a web interface, that will be capable of performing
a OAUTH dance with a third party OIDC provider.

The same web server, could be used for generating and perhaps
interacting with GDS CLI when it comes to a credential generation and
management.

As a bonus, we're adding a healthcheck endpoint capable of telling us if
we're keeping the connection with redis alive. Could prove itself useful
on Kubernetes :troll: